perf(app-router): isolate action request detection
